stocks-service/requirements.txt
yliveticker==0.3.0
websockets==10.0
yfinance==0.1.64
exponent_server_sdk==2.0.0
pyfcm==1.5.4
azure-storage-blob==12.9.0
python-dotenv==0.19.2
keras==2.6.0
opencv-python==4.5.4.60
tensorflow==2.6.0
tensorflow-gpu==2.6.0